test(search): split repair regression fixture
