+++ internal/repository/search_repair_integration_test.go
@@ -205,63 +205,6 @@ func TestSearchRepairConvergesWhitespaceDelimitedCanonicalEvidence(t *testing.T)
 	require.Zero(t, projection.DriftedDocuments)
 }
 
-func TestSearchRepairUpdatesExistingDocumentUsingStoredFenceValues(t *testing.T) {
-	adminDB, appDB, rls, cleanup := setupLedgerRepositoryDB(t)
-	defer cleanup()
-	ctx := context.Background()
-	teamID := createLedgerTeam(t, adminDB, rls, "search-repair-stored-fence-team")
-	ownerID := createLedgerProfile(t, adminDB, rls, teamID, "search-repair-stored-fence-owner")
-	insertSearchTestContract(t, adminDB, rls, "search-repair-stored-fence", 2, "exact", "")
-	repo := NewSearchRepository(appDB, rls)
-	ledger := NewLedgerRepository(appDB, rls)
-	content := "canonical stored fence evidence"
-	ingest := createSearchRepairAcceptedIngest(t, ctx, ledger, CreateIngestInput{
-		TeamID: teamID, OwnerProfileID: ownerID, IdempotencyKey: "search-repair-stored-fence-evidence",
-		RequestHash: sha256Hex(content), Evidence: []EvidenceInput{{Content: content}},
-	})
-	document, err := repo.UpsertSearchDocument(ctx, UpsertSearchDocumentInput{
-		TeamID: teamID, OwnerProfileID: ownerID, SourceKind: "evidence", SourceID: ingest.Evidence[0].FragmentID,
-		SourceVersion: 1, DocumentText: "stale stored fence evidence",
-	})
-	require.NoError(t, err)
-	contract, err := repo.GetActiveSearchContract(ctx)
-	require.NoError(t, err)
-	run, claimed, err := repo.ReserveSearchRepairRun(ctx, SearchRepairRunInput{
-		EmbeddingContractID: contract.EmbeddingContractID, EmbeddingDimensions: contract.EmbeddingDimensions,
-		LocalRunDate: time.Now().UTC(), CreateIfMissing: true, WorkerID: "search-repair-stored-fence-worker", Lease: time.Minute,
-	})
-	require.NoError(t, err)
-	require.True(t, claimed)
-	selected, more, err := repo.SelectSearchRepairDocuments(ctx, SearchRepairSelectionInput{
-		EmbeddingContractID: contract.EmbeddingContractID, EmbeddingDimensions: contract.EmbeddingDimensions, Limit: 1,
-	})
-	require.NoError(t, err)
-	require.False(t, more)
-	require.Len(t, selected, 1)
-	require.Equal(t, document.SearchDocumentID, selected[0].SearchDocumentID)
-	require.Equal(t, content, selected[0].DocumentText)
-
-	apply, err := repo.ApplySearchRepair(ctx, ApplySearchRepairInput{
-		RunID: run.RunID, LeaseToken: run.LeaseToken,
-		EmbeddingContractID: contract.EmbeddingContractID, EmbeddingDimensions: contract.EmbeddingDimensions,
-		SearchIndexGenerationID: contract.SearchIndexGenerationID, IndexGeneration: contract.IndexGeneration,
-		Documents: []SearchRepairEmbedding{{SearchRepairDocument: selected[0], Embedding: []float32{0.5, 0.5}}},
-	})
-	require.NoError(t, err)
-	require.EqualValues(t, 1, apply.UpdatedCount)
-	require.Zero(t, apply.RemainingDriftedCount)
-
-	var repairedText string
-	require.NoError(t, rls.WithSystemTx(ctx, adminDB, func(tx *gorm.DB) error {
-		return tx.Raw(`
-			SELECT document_text
-			FROM search_documents
-			WHERE team_id = ?::uuid AND search_document_id = ?::uuid
-		`, teamID, document.SearchDocumentID).Row().Scan(&repairedText)
-	}))
-	require.Equal(t, content, repairedText)
-}
-
 func TestSearchRepairRepairsCanonicalEvidenceOwnerMismatch(t *testing.T) {
 	adminDB, appDB, rls, cleanup := setupLedgerRepositoryDB(t)
 	defer cleanup()
